refactor: create components for app action card and result and update types
